splitted system_clock in differnet files
